refactor(ksize): file refactor, add exit if no ai mutated text, add prompt to repo
