+++ README.md
@@ -1,2 +1,2 @@
-# ksize_catalog_parser
-Product catalog parser
+# distributors_catalog_parser
+Product catalog parser for ksize and [car-smart](https://car-smart.ru/)
