--- parser/_common.py
@@ -1,3 +1,0 @@
-import os
-
-KSIZE_URL = os.getenv("KSIZE_URL")
