+++ src/ksize/parse.py
@@ -6,7 +6,7 @@ logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(message)s")
 logging.getLogger("httpx").setLevel(logging.WARNING)
 
 import os
-from parser import process_caregory
+from src.ksize.parser import process_caregory
 
 process_caregory(os.getenv("SUBCATEGORY_URL"))
 logging.info("Parsing completed")
