+++ src/ksize/parser/_common.py
@@ -1,0 +1,4 @@
+import os
+
+KSIZE_URL = os.getenv("KSIZE_URL")
+DESCRIPTION_PROMPT="Оптимизируй текст под SEO для каждого товара. Основной товар - автомагнитолы. Исключи серию и модель магнитолы. Задействуй не более 500 символов. Не дублируй характеристики в краткое описание. Добавь 10 характеристик в виде таблицы с использованием тегов ul и li"
