+++ src/ksize/parser/categories.py
