+++ src/ksize/parser/products/__init__.py
@@ -1,2 +1,2 @@
 from .get import get_product, get_products
-from .model import Product
+from .schemas import Product
