+++ src/ksize/parser/products/schemas.py
